UCS-76 : rename PipChMessage to AttributeChangeMessage
