+++ UsageControlFramework/src/main/java/it/cnr/iit/usagecontrolframework/rest/UsageControlFramework.java
@@ -34,9 +34,9 @@ import it.cnr.iit.ucs.properties.components.PepProperties;
 import it.cnr.iit.ucs.properties.components.PipProperties;
 import it.cnr.iit.ucsinterface.contexthandler.AbstractContextHandler;
 import it.cnr.iit.ucsinterface.forwardingqueue.ForwardingQueue;
+import it.cnr.iit.ucsinterface.message.attributechange.AttributeChangeMessage;
 import it.cnr.iit.ucsinterface.message.endaccess.EndAccessMessage;
 import it.cnr.iit.ucsinterface.message.endaccess.EndAccessResponse;
-import it.cnr.iit.ucsinterface.message.pipch.PipChMessage;
 import it.cnr.iit.ucsinterface.message.reevaluation.ReevaluationMessage;
 import it.cnr.iit.ucsinterface.message.reevaluation.ReevaluationResponse;
 import it.cnr.iit.ucsinterface.message.startaccess.StartAccessMessage;
@@ -294,7 +294,7 @@ public class UsageControlFramework implements UCSInterface {
 
     @Override
     @Async
-    public void retrieveRemote( PipChMessage messagePipCh ) {
+    public void retrieveRemote( AttributeChangeMessage messagePipCh ) {
         // TODO check if sent
         requestManager.sendMessageToCH( messagePipCh );
     }
